refactor: Move content of the queries file to queries folder
