+++ src/components/Papers/HarvestBanner.jsx
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types'
 import React from 'react'
-import { buildTriggersQueryByKonnectorSlug } from 'src/helpers/queries'
+import { buildTriggersQueryByKonnectorSlug } from 'src/queries'
 
 import { useQuery, isQueryLoading } from 'cozy-client'
 import { LaunchTriggerCard } from 'cozy-harvest-lib'
